add FF config

+++ tools/data_converter/scannet_data_utils.py
@@ -609,7 +609,7 @@ class ScanNetSVData(object):
 
             pts_filename = osp.join(self.split_dir,
                                     f'{sample_idx}_pc.npy')
-            points = np.load(pts_filename)
+            points = np.load(pts_filename).astype(np.float32)
             mmcv.mkdir_or_exist(osp.join(self.save_path, 'points'))
             points.tofile(
                 osp.join(self.save_path, 'points', f'{sample_idx}.bin'))
